Explicitly cast the v6502_signedValueOfByte return value
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/v6502/trunk@753 96dfbd92-d197-e211-9ca9-00110a534b34

+++ v6502/mem.c
@@ -144,7 +144,7 @@ int8_t v6502_signedValueOfByte(uint8_t byte) {
 //	if (byte & 0x80) {
 //		return (0 - (BYTE_MAX - byte));
 //	}
-	return byte;
+	return (int8_t)byte;
 }
 
 uint8_t v6502_byteValueOfSigned(int8_t i) {
